pyproject.toml
[project]
name = "code-review-mcp"
version = "0.1.0"
description = "MCP server for AI-powered code review using LiteLLM and an LLM proxy"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "mcp>=1.1.0",
    "litellm>=1.0.0",
    "pydantic>=2.0.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0.0"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.pytest.ini_options]
pythonpath = ["src"]
